+++ src/burn/drv/taito/d_flstory.cpp
@@ -1,5 +1,6 @@
 #include "tiles_generic.h"
 #include "taito_m68705.h"
+#include "zet.h"
 #include "driver.h"
 #include "dac.h"
 extern "C" {
@@ -669,7 +670,7 @@ UINT8 __fastcall flstory_sound_read(UINT16 address)
 
 static INT32 flstoryDACSync()
 {
-	return (float)(nBurnSoundLen * (ZetTotalCyclesPrecise() / (4000000.000 / (nBurnFPS / 100.000))));
+	return (INT32)(float)(nBurnSoundLen * (ZetTotalCyclesPrecise() / (4000000.000 / (nBurnFPS / 100.000))));
 }
 
 static INT32 DrvDoReset()
